Use http crate, rename query module, and remove the prelude
