+++ src/lib.rs
@@ -1,13 +1,13 @@
-pub mod api;
-pub mod client;
-pub mod endpoint;
-pub mod error;
-pub mod http;
-pub mod prelude;
-pub mod query;
-pub mod tmdb;
+mod api;
+mod client;
+mod endpoint;
+mod error;
+mod query_parameters;
+mod tmdb;
 
+pub use api::*;
 pub use client::Client;
+pub use endpoint::Endpoint;
 pub use error::Error;
-pub use query::QueryParameters;
+pub use query_parameters::QueryParameters;
 pub use tmdb::Tmdb;
