--- src/prelude.rs
@@ -1,3 +1,0 @@
-pub use crate::client::Client;
-pub use crate::endpoint::Endpoint;
-pub use crate::tmdb::Tmdb;
